Digest.pm - version checks should not have quoted versions
In 5.39.1 passing an argument to a missing import is an error.
When the version is quoted it is treated as an argument to
import(), but perl special cases import so when it is missing
it silently ignores the call entirely. (Exporter::import would
do a version check.)

Simply removing the quotes makes 5.39.1 happy.